Treat KaitaiStructError the same as EOFError (KS 0.8-compatible)
This fixes for example the ksv's inconsistency in handling EOFError (for
which you get a message box modal "window" that you can close and
continue) and validation errors (which crash ksv with a stack trace).

A special helper class `KSErrorMatcher` was needed because
`KaitaiStructError` might not exist in the runtime library (which is the
case in KS 0.8 and earlier).