ksv, ksdump: use `__dir__` (not `$0`) to get script dir more reliably
In most cases, the `$0` variable (note: the `$PROGRAM_NAME` used in the
old code is an alias of `$0` in Ruby) works fine. However, for some
unusual ways of launching our scripts, it doesn't because it essentially
asks the wrong question: we don't really care what was the entry point
of the current process (which is what `$0` means), we only want to get
the directory of the current script so we can resolve relative paths
from it.

In particular, the old way didn't work when I launched `bin/ksdump` via
https://github.com/ruby-prof/ruby-prof in order to profile it (for
example as `ruby-prof -f ruby_prof.log bin/ksdump -- -f json sample.bin
format.ksy > sample.json`), because `$0` in my case was
`"C:/Ruby32-x64/bin/ruby-prof"`, which is obviously not what
`bin/ksdump` expected.