Simplify enumeration of struct properties
There's no point in calling `public_methods` with no arguments first
(this method is defined as `public_methods(all=true)`, so the `all`
parameter is true by default), which exactly includes methods from *all*
ancestor classes, only to remove all methods from ancestor classes
manually using `Set` subtraction afterwards. It's much easier to set the
`all` parameter to false in the first place, which will directly get us
what we want.

Not to mention that after doing some profiling with
https://github.com/ruby-prof/ruby-prof of `bin/ksdump` on some big file,
the `Set` manipulation was apparently quite slow as well, which really
adds up for hundreds of millions of objects, for example. Note that even
though a .ksy file typically only defines a handful of types, so there
is only a handful of distinct classes in the generated parser module and
the set of methods is the same for all objects of each class, the code
in ksv doesn't take advantage of this and always enumerates methods for
every object it sees as if it hadn't seen any object of the same class
before. This is fine for simplicity and we don't have any evidence that
"caching" of the set of methods per class would help, but the point is
that you ideally want to avoid slow ways of doing things there, which
the old way of enumerating methods using `Set` subtraction was.