Revert "Support forward slashes in .ksy path on Windows"
As explained in
https://github.com/kaitai-io/kaitai_struct/issues/507#issuecomment-1679530418,
retrying the path with all forward slashes `/` replaced with backslashes
`\` was a band-aid fix, not a proper solution. It's better to have the
compiler preserve input .ksy paths exactly in JSON keys so that
applications like `ksv` don't have to do any trial-and-error or complex
path resolution. The compiler has already been changed to do that in
https://github.com/kaitai-io/kaitai_struct_compiler/commit/4c5096ba53d559fcfe235a5a0df1082e7f427e49,
so this `ksv` hack can be reverted.

This (partially) reverts commit d9a3dbdc6c06ffd47c94e9f012802b0d61b68af9.