Revert "parser_spec: add partial_array_int + partial_array_struct tests"
This reverts commit 231beb54b77d0161799d5227865f4d6f7fcd56f8.

This test currently fails with KSC 0.10, which CI uses. This is bad
because it blocks the CI pipeline from continuing. Furthermore, I don't
understand why this test case is included in the visualizer test suite
when it only tests the compiler (this is apparently the case for all
tests in `spec/parser_spec.rb`, but at least the others pass). And we
already have a very similar test in the main KST suite, see
https://github.com/kaitai-io/kaitai_struct_tests/blob/0914597c9256b19259c194bfa8877af2218152e6/formats/debug_array_user_eof_exception.ksy