.gemspec: add an upper bound on the version of `benchmark`
Fixes the warning from `gem build`:

```
$ gem build kaitai-struct-visualizer.gemspec
WARNING:  open-ended dependency on benchmark (>= 0.1.0) is not recommended
  if benchmark is semantically versioned, use:
    add_runtime_dependency "benchmark", "~> 0.1", ">= 0.1.0"
WARNING:  See https://guides.rubygems.org/specification-reference/ for help
```

Note: https://rubygems.org/gems/benchmark/versions/0.4.1 is the latest
version as of July 2025. We're assuming semantic versioning, so all
`< 0.5.0` versions should be compatible, but we can't know whether
`0.5.0` will be, so we exclude it.