.rubocop.yml: do not override the default `AllCops/Exclude`
It seems that the reason RuboCop was always hanging in CI was that it
was stuck scanning the entirety of the massive `./vendor/bundle`
directory with over 2000 files where the
https://github.com/ruby/setup-ruby action stores the gems installed
using `bundle install`. By default, that wouldn't happen because
`vendor/**/*` is in the default `AllCops/Exclude` list (see
https://github.com/rubocop/rubocop/blob/v1.79.1/config/default.yml#L68),
but we overrode it with our own array.

It is possible to set up
[inheritance](https://docs.rubocop.org/rubocop/1.79/configuration.html#merging-arrays-using-inherit_mode)
so that the `Exclude` arrays get merged, but for now it's easier to just
remove our custom `Exclude` array, since there are no
`spec/compiled/*.rb` files in the repo yet.